Version 0.6.1 New estimate for mad tail
Changed the convolution test to adjust for the MAD from the tail.
Distribution function for cdf amended to do upper tail as well as lower.
Added distribution function for pdf.
Adjusted the size of the convolution array so that it's a multiple of small primes.
Set minimum size of convolution array to 1,000,000 and split the extra points between extending the range and increasing the density.
Added the maximum size, m, to the output file name.
Added additional instrumentation to the trace data.